+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 		tab) in full color, darken all others. When inverted, show in full color
 		branches with a Strahler number strictly lower than the specified one.
 
+- The Z plane display now supports stack viewer layers. If multiple layers are
+  shown in a stack viewer, its Z plane will render all visible layers on top of
+  each other.
 
 Review widget: